some improves

+++ include/uapi/types.h
@@ -12,7 +12,7 @@ typedef int64_t  off_t;
 typedef uint64_t dev_t;     // 장치 식별자
 typedef uint64_t ino_t;     // Inode 번호
 typedef uint32_t mode_t;    // 파일 속성
-typedef uint32_t nlink_t;   // 하드 링크 수
+typedef uint64_t nlink_t;   // 하드 링크 수
 typedef uint32_t uid_t;     // User ID
 typedef uint32_t gid_t;     // Group ID
 typedef uint64_t blksize_t; // 블록 크기
